[libc++] Clean up the no-unicode CI job
It was added after we changed the way the CI jobs are run, in particular
how they are pinned down to Linux instances only. As a result, the job
would sometimes run on Mac machines, which we're trying to keep only for
jobs that absolutely need it due to capacity concerns.

NOKEYCHECK=True
GitOrigin-RevId: 8660b89c0c2cccfc5b8ecb70918a3aa8ddcd67bb